pyproject.toml
[tool.poetry]
name = "deckard"
version = "0.1.0"
description = ""
authors = ["UNB Libraries <libsystems@unb.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "4.12.3"
dictdiffer = "0.9.0"
flask = "3.0.2"
html2text = "2020.1.16"
huggingface-hub = "0.20.3"
lancedb = "0.5.6"
langchain = "0.1.8"
langchain-community = "0.0.21"
llama-cpp-python = "0.2.47"
pandas = "2.2.0"
psutil = "5.9.8"
python-box = "7.1.1"
pyyaml = "6.0.1"
requests = "2.31.0"
sentence_transformers = "2.3.1"
slack-bolt = "1.18.1"
slack-sdk = "3.27.0"

[tool.poetry.scripts]
"api:start" = "deckard.interfaces.api:start"
"build:rag" = "deckard.interfaces.ragbuild:start"
"evaluate" = "deckard.interfaces.ragevaluate:start"
"query:llm" = "deckard.interfaces.llmdirectquery:query"
"query:rag" = "deckard.interfaces.ragquery:query"
"search:embeddings" = "deckard.interfaces.embeddingsearch:search"
"slackbot:start" = "deckard.interfaces.slackbot:start"
"test" = "deckard.interfaces.test:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
